Suggesting 3 users so that calendar can fit
Using a small screen size and with a number of 5 suggestions,
you have to scrool untill all posts finish, and then you can see the
calendar properly on the right side. By now, we avoid this changing the
number of suggestion, but we should make the right side scrollable
Twitter style.