Fixed code due to first test.
The bug came from the fact that a like can be deleted, but the
notification is still existing. The SQL join statement generated 0 rows
and the array accessing by ids gave exceptions.
Now everything is fixed, and even if a like is deleted, the notification
will be still delivered (as Instagram does).