Made notifications improvements
Now all the unread notifications are on the top of the modal, ordered by
date. Also, the notifications modal is created once this file is loaded,
and refresh if only new notifications are found.
The modal will be empty if only no notifications are found for the
current user (before, the modal was empty if no NEW notifications were
found).